docshell/test/mochitest/test_bug660404.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=660404
-->
<head>
  <title>Test for Bug 660404</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=660404">Mozilla Bug 660404</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 660404 **/
SimpleTest.waitForExplicitFinish();

var w;

function continueTest() {
  // Do this async so the load event can finish firing
  SimpleTest.executeSoon(function() {
    w.onpagehide = function(ev) {
      is(ev.persisted, true, "Should be bfcached when navigating to multipart");
    };
    w.location.href = "file_bug660404";
  });
}

function finishTest() {
  is(w.document.documentElement.textContent, "opener.finishTest();");
  is(w.document.documentElement.innerHTML, "<head><script>opener.finishTest();</" +
                                           "script></head>");
  w.close();
  SimpleTest.finish();
}

// Have to open a new window, since there's no bfcache in subframes
w = window.open("file_bug660404-1.html");

</script>
</pre>
</body>
</html>
